Updated folder structure for static resources to include a dedicated Prism directory, as well as a themes subdirectory
This is all a bit of a prefactor for eventually having other static resource files, and possibly supporting other Prism themes in the future
