+++ .prettierignore
@@ -22,7 +22,6 @@ yarn.lock
 
 # PrismJS - these files are automatically updated & minified,
 # so no need for prettier to format them.
-nebula-logger/core/main/log-management/staticresources/LoggerResources/prism.css
-nebula-logger/core/main/log-management/staticresources/LoggerResources/prism.min.css
 nebula-logger/core/main/log-management/staticresources/LoggerResources/prism.js
 nebula-logger/core/main/log-management/staticresources/LoggerResources/prism.min.js
+nebula-logger/core/main/log-management/staticresources/LoggerResources/themes/*.*
